# Conflicts:
#	server/src/main/java/spring/manager/ItemManager.java
#	server/src/main/java/spring/manager/OrderManager.java

+++ server/src/main/java/spring/repositories/ItemCategoryRepository.java
@@ -3,8 +3,9 @@ package spring.repositories;
 import org.springframework.data.repository.PagingAndSortingRepository;
 import spring.entity.ItemCategories;
 
+import java.math.BigInteger;
 import java.util.Optional;
 
-public interface ItemCategoryRepository extends PagingAndSortingRepository<ItemCategories,Long> {
-    Optional<ItemCategories> findById( Long id );
+public interface ItemCategoryRepository extends PagingAndSortingRepository<ItemCategories,BigInteger> {
+    Optional<ItemCategories> findById( BigInteger id );
 }
